requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.4.1
    # via django
django==3.2.5
    # via
    #   -r requirements.in
    #   django-filter
    #   django-redis
    #   djangorestframework
django-filter==2.4.0
    # via -r requirements.in
django-redis==5.0.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
markdown==3.3.4
    # via -r requirements.in
pytz==2021.1
    # via django
redis==3.5.3
    # via django-redis
sqlparse==0.4.1
    # via django
